+++ src/main/resources/META-INF/services/org.openapitools.codegen.CodegenConfig
@@ -1,1 +1,1 @@
-com.tweesky.cloudtools.codegen.PostmanV2Generator
+com.adyen.codegen.postman.PostmanV2Generator
